requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
coloredlogs==15.0.1
exceptiongroup==1.1.2
fastapi==0.100.0
filelock==3.12.2
flatbuffers==23.5.26
frozenlist==1.3.3
fsspec==2023.6.0
h11==0.14.0
humanfriendly==10.0
idna==3.4
Jinja2==3.1.2
joblib==1.3.1
lightning-utilities==0.9.0
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.25.1
onnxruntime==1.15.1
packaging==23.1
pandas==2.0.3
protobuf==4.23.4
psycopg2-binary==2.9.6
pydantic==2.0.2
pydantic_core==2.1.2
python-dateutil==2.8.2
pytorch-lightning==2.0.5
pytz==2023.3
PyYAML==6.0
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
sympy==1.12
threadpoolctl==3.1.0
torch==2.0.1
torchmetrics==1.0.0
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
yarl==1.9.2
